General class for optimizers with extrapolation
